Top 3 class emojis per fight

+++ Modules/JsonObjects/xivdbCharacter/XIVDBCharacter.cs
@@ -11,14 +11,14 @@ namespace XivDB {
         public string last_synced { get; set; }
         public string data_last_changed { get; set; }
         public string data_hash { get; set; }
-        public int update_count { get; set; }
+        public double update_count { get; set; }
         public string achievements_last_updated { get; set; }
         public string achievements_last_changed { get; set; }
-        public int achievements_public { get; set; }
-        public int achievements_score_reborn { get; set; }
-        public int achievements_score_legacy { get; set; }
-        public int achievements_score_reborn_total { get; set; }
-        public int achievements_score_legacy_total { get; set; }
+        public double achievements_public { get; set; }
+        public double achievements_score_reborn { get; set; }
+        public double achievements_score_legacy { get; set; }
+        public double achievements_score_reborn_total { get; set; }
+        public double achievements_score_legacy_total { get; set; }
         public string deleted { get; set; }
         public int priority { get; set; }
         public int patch { get; set; }
@@ -32,8 +32,8 @@ namespace XivDB {
         public string url_xivdb { get; set; }
         public string url_lodestone { get; set; }
         public string url_type { get; set; }
-        public float achievements_score_reborn_percent { get; set; }
-        public int achievements_score_legacy_percent { get; set; }
+        public double achievements_score_reborn_percent { get; set; }
+        public double achievements_score_legacy_percent { get; set; }
         public Extras extras { get; set; }
     }
 }
